Move checkRccSegmentApplicability function into root lib folder
