+++ packages/core/lib/checkRccSegmentApplicability.test.ts
@@ -1,6 +1,6 @@
-import type { Offence, Result } from "../../../../types/AnnotatedHearingOutcome"
+import type { Offence, Result } from "../types/AnnotatedHearingOutcome"
 
-import ResultClass from "../../../../types/ResultClass"
+import ResultClass from "../types/ResultClass"
 import checkRccSegmentApplicability, { RccSegmentApplicability } from "./checkRccSegmentApplicability"
 
 describe("checkRccSegmentApplicability", () => {
